+++ src/main/java/com/rollingpaper/ggeujeogggeujeog/paper/application/PaperServiceImpl.java
@@ -10,7 +10,7 @@ import org.springframework.web.multipart.MultipartFile;
 import com.rollingpaper.ggeujeogggeujeog.paper.domain.Paper;
 import com.rollingpaper.ggeujeogggeujeog.paper.domain.PaperRepository;
 import com.rollingpaper.ggeujeogggeujeog.paper.exception.NoSuchPaperException;
-import com.rollingpaper.ggeujeogggeujeog.paper.infrastructure.ImageStorage;
+import com.rollingpaper.ggeujeogggeujeog.paper.domain.ImageStorage;
 import com.rollingpaper.ggeujeogggeujeog.paper.presentation.dto.PaperResponseDto;
 import com.rollingpaper.ggeujeogggeujeog.paper.presentation.dto.PaperUpdateRequestDto;
 import com.rollingpaper.ggeujeogggeujeog.paper.presentation.dto.PaperWriteRequest;
