+++ src/main/java/com/rollingpaper/ggeujeogggeujeog/paper/domain/ImageStorage.java
@@ -1,4 +1,4 @@
-package com.rollingpaper.ggeujeogggeujeog.paper.infrastructure;
+package com.rollingpaper.ggeujeogggeujeog.paper.domain;
 
 import java.io.IOException;
 
@@ -7,6 +7,8 @@ import org.springframework.web.multipart.MultipartFile;
 public interface ImageStorage {
 
 	default void init() throws IOException {}
+
 	String store(MultipartFile imageFile);
+
 	boolean doesImageFileExist(String imageFileName);
 }
