+++ src/main/java/com/rollingpaper/ggeujeogggeujeog/paper/infrastructure/NcpObjectStorage.java
@@ -9,6 +9,7 @@ import org.springframework.web.multipart.MultipartFile;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.rollingpaper.ggeujeogggeujeog.common.util.FileNameUtil;
+import com.rollingpaper.ggeujeogggeujeog.paper.domain.ImageStorage;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
